Choose encryption algoritm based on cert name
In [this][1] article the author presents a way to have Nginx work
with both RSA and ECC keys at the same time. This is achieved by
just including two `ssl_certificate[_key]` lines which points to
different files. By then ordering the `ssl_ciphers` in your
preferred way you can have Nginx automatically use the correct RSA
or ECDSA key/certificate.

In the article the author add ".ecc" to the files using ECDSA (ECC
is underlying technology), and ".rsa" to those using RSA. So to handle
this in our script we have a regex that checks if it can match once of
these alternatives.
All of these may appear anywhere in the `cert-name`:

    -rsa
    .rsa
    -ecc
    .ecc
    -ecdsa
    .ecdsa

All of these are also case insensitive. Having these somewhere in the
name will override the `USE_ECDSA` environment variable, so special
cases like this can be made. If the "rsa" string is matched an RSA key
is created, while the others will be of the ECDSA type.

[1]: https://medium.com/hackernoon/rsa-and-ecdsa-hybrid-nginx-setup-with-letsencrypt-certificates-ee422695d7d3